% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{download_image_files}
\alias{download_image_files}
\title{Download image files}
\usage{
download_image_files(
  version = "current",
  destination = getwd(),
  datasets = NULL,
  max_results = NULL,
  overwrite = F
)
}
\arguments{
\item{version}{(Optional) Version of the dataset to reference (defaults to
"current"), can also be "next" for unreleased versions.}

\item{destination}{directory to write files to, defaults to working directory}

\item{datasets}{(Optional) Character vector of dataset IDs to include
(default to all datasets).}

\item{max_results}{(Optional) Max number of records to load for each table (defaults to
entire table).}
}
\description{
Download image files
}
