% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_annotated_messages}
\alias{get_annotated_messages}
\title{Get message annotations}
\usage{
get_annotated_messages(
  version = "current",
  datasets = NULL,
  max_results = NULL
)
}
\arguments{
\item{version}{(Optional) Version of the dataset to reference (defaults to
"current"), can also be "next" for unreleased versions.}

\item{datasets}{(Optional) Character vector of dataset IDs to include
(default to all datasets).}

\item{max_results}{(Optional) Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get message annotations
}
