% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_current_version}
\alias{get_current_version}
\title{Returns the version tag for the current version of refbank}
\usage{
get_current_version()
}
\description{
Returns the version tag for the current version of refbank
}
