% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets}
\usage{
get_datasets(max_results = NULL)
}
\arguments{
\item{max_results}{(Optional) Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get datasets
}
