% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_messages}
\alias{get_messages}
\title{Get messages}
\usage{
get_messages(
  version = "current",
  datasets = NULL,
  max_results = NULL,
  include_trial_data = F,
  include_player_data = F,
  include_image_data = F,
  include_condition_data = F
)
}
\arguments{
\item{version}{(Optional) Version of the dataset to reference (defaults to
"current"), can also be "next" for unreleased versions.}

\item{datasets}{(Optional) Character vector of dataset IDs to include
(default to all datasets).}

\item{max_results}{(Optional) Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get messages
}
