% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{refbank}
\alias{refbank}
\title{Reference to Refbank Redivis dataset}
\usage{
refbank(version)
}
\arguments{
\item{version}{(Optional) Version of the dataset to reference (defaults to
"current"), can also be "next" for unreleased versions.}
}
\description{
Reference to Refbank Redivis dataset
}
