% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{refbank_derived}
\alias{refbank_derived}
\title{Reference to Refbank Redivis workflow for processed outputs}
\usage{
refbank_derived()
}
\description{
Reference to Refbank Redivis workflow for processed outputs
}
