% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refbankr-package.R
\docType{package}
\name{refbankr-package}
\alias{refbankr}
\alias{refbankr-package}
\title{refbankr: Downloads Data From Refbank}
\description{
Package for downloading data from refbank repository of reference game data.
}
\author{
\strong{Maintainer}: Michael Frank \email{mcfrank@stanford.edu}

Authors:
\itemize{
  \item Michael Frank \email{mcfrank@stanford.edu}
}

}
\keyword{internal}
