% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_cosine_similarities}
\alias{get_cosine_similarities}
\title{Get cosine similarities of sbert embeddings}
\usage{
get_cosine_similarities(datasets = NULL, type = NULL, max_results = NULL)
}
\arguments{
\item{datasets}{(Optional) Character vector of dataset IDs to include
(default to all datasets).}

\item{type}{(Optional) Character vector of which similarity comparisons to return
options are "to_last", "to_next", "to_first", "diverge", "diff", "idiosyncrasy"
(default to all comparisons).}

\item{max_results}{(Optional) Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get cosine similarities of sbert embeddings
}
