% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_messages}
\alias{get_messages}
\title{Get messages}
\usage{
get_messages(datasets = NULL, max_results = NULL)
}
\arguments{
\item{datasets}{(Optional) Character vector of dataset IDs to include
(default to all datasets).}

\item{max_results}{(Optional) Max number of records to load for each table (defaults to
entire table).}
}
\description{
Get messages
}
